chore(deps): update npm, cargo, and CI dependencies (#59)
Major bumps (verified — build + full test suites pass):
- rusqlite 0.31 -> 0.40: read COUNT(*) as i64 then cast to usize at the
  two tray-count sites (lib.rs, converter.rs); the 0.38 change disabled the
  usize/u64 FromSql impls by default.
- typescript 5.8 -> 6.0: add `types: ["node"]` to tsconfig — 6.0 defaults
  `types` to [], which otherwise drops @types/node and breaks the `node:*`
  imports in the release-script test.
- @types/node 25 -> 26
- actions/checkout v6.0.3 -> v7.0.0 (both already node24; no behaviour change)
- tauri-apps/tauri-action v0.6.2 -> v1.0.0 (all inputs we pass unchanged)

Minor/in-range bumps:
- vite 8.0.16 -> 8.1.3, @vitejs/plugin-react 6.0.2 -> 6.0.3,
  @tauri-apps/cli 2.11.2 -> 2.11.4
- cargo update: tauri 2.11.3 -> 2.11.5, uuid 1.23.3 -> 1.23.4, transitive deps

+++ src-tauri/src/converter.rs
@@ -339,9 +339,9 @@ fn process_queue(app: &AppHandle, db: &Arc<Mutex<Connection>>, converter: &Conve
             db.query_row(
                 "SELECT COUNT(*) FROM jobs WHERE status = 'queued'",
                 [],
-                |row| row.get::<_, usize>(0),
+                |row| row.get::<_, i64>(0),
             )
-            .unwrap_or(0)
+            .unwrap_or(0) as usize
         };
 
         let _ = app.emit(
